[Refactor/#31] 긴 Style, Widget 분리
