+++ dnf_cal/lib/widgets/MainPage/CalendarStyle.dart/BuildDaysOfWeekStyle.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:table_calendar/table_calendar.dart';
+
+DaysOfWeekStyle buildDaysOfWeekStyle() {
+  return DaysOfWeekStyle(
+    decoration: BoxDecoration(
+      color: Color(0xff0f172f),
+    ),
+    weekdayStyle: TextStyle(
+      color: Colors.white,
+      fontSize: 12,
+      fontFamily: 'DNFForgedBlade',
+      fontWeight: FontWeight.w300,
+    ),
+    weekendStyle: TextStyle(
+      color: Colors.white,
+      fontSize: 12,
+      fontFamily: 'DNFForgedBlade',
+      fontWeight: FontWeight.w300,
+    ),
+  );
+}
